fix: 3.0 lts 清零 【背景】 3.0 lts 告警清零
【修改方案】
添加强制类型转换

【影响】
对现有的产品编译不会有影响。

re #I4VX5D

+++ tzdriver/src/smc_smp.c
@@ -317,7 +317,7 @@ static void ShowInBitmap(int *cmdIn, uint32_t len)
     for (idx = 0; idx < MAX_SMC_CMD; idx++) {
         if (HmTestBit(idx, g_cmdData->inBitmap)) {
             bitmap[idx] = '1';
-            cmdIn[in++] = idx;
+            cmdIn[in++] = (int)idx;
         } else {
             bitmap[idx] = '0';
         }
